+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/App.kt
@@ -8,7 +8,7 @@ import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import io.music_assistant.client.api.ServiceClient
-import io.music_assistant.client.ui.compose.nav.NavigationRoot
+import io.music_assistant.client.ui.compose.nav.TopLevelNavRoot
 import io.music_assistant.client.ui.theme.AppTheme
 import io.music_assistant.client.ui.theme.SystemAppearance
 import io.music_assistant.client.ui.theme.ThemeSetting
@@ -49,6 +49,6 @@ fun App() {
     }
     SystemAppearance(isDarkTheme = darkTheme)
     AppTheme(darkTheme = darkTheme) {
-        NavigationRoot()
+        TopLevelNavRoot()
     }
 }
